Adding subroutine to calculate div(u)
This is based on the temperature field and is used to determine
the velocity divergence in the variable-coefficient pressure-Poisson
equation.
Also, this is the divergence needed by density equation in next
timestep.

Computes div(u) in Z pencil

As side effect updates density and temperature arrays with the new
values.