Adding temperature field to restart variables.
This allows pressure0 (the thermodynamic pressure) to be determined
in a restart.

Note this is a sub-optimal solution as it requires knowing the EOS,
a better approach would be to write/read pressure0, the temperature
field is then subsequently computed by eos() in calc_divu() but I
don't know how to write a single value, this is thus the simplest,
if least elegant solution.

Writing pressure0 also will have the advantage of making restart
files smaller: NX*NY*NZ + 1 vs 2 * NX*NY*NZ