Bug fix - improves performance of variable-coefficient Poisson solver
In divergence_corr() the full set of residuals are computed, even
if convergence has occurred and as a result the array pp3corr which
contains the pressure is overwritten. The pressure gradients are
still correct however, but we end up back at the zero pressure field
(residual->0 at convergence) and therefore as simulation goes on we
have to work harder to get to new pressure field. By storing the current
pressure field (as was intended), we do not need to start again from zero
at each timestep.